requirements.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.15
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.9.2
fonttools==4.55.4
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
limiter==0.5.0
limits==3.14.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
numpy==2.2.2
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rich==13.9.4
six==1.17.0
StrEnum==0.4.15
stripe==11.3.0
timedelta==2020.12.3
token-bucket==0.3.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0